replaced/cast size_t by/to unsigned int
git-svn-id: file:///Users/gsalam/svn/fastjet/trunk/fastjet-release@2704 9ab96302-e443-4389-8a14-f726967065a6

+++ example/04-constituents.cc
@@ -107,9 +107,9 @@ int main(){
     // get the constituents of the jet
     vector<fastjet::PseudoJet> constituents = inclusive_jets[i].constituents();
 
-    printf("%5u %15.8f %15.8f %15.8f %8lu\n",
+    printf("%5u %15.8f %15.8f %15.8f %8u\n",
 	   i, inclusive_jets[i].rap(), inclusive_jets[i].phi(),
-	   inclusive_jets[i].perp(), constituents.size());
+	   inclusive_jets[i].perp(), (unsigned int) constituents.size());
 
     printf("       ");
     for (unsigned int j=0; j<constituents.size(); j++){
